+++ R/assortativity.R
@@ -14,7 +14,7 @@
 #' \dontrun{
 #' # create a mixing matrix of the demonstration network based on vertex 
 #' # categorical attribute for political stance "vosonCA_Stance"
-#' g <- loadDemoGraph("DividedTheyBlog_40Alist_release.graphml")
+#' g <- loadPackageGraph("DividedTheyBlog_40Alist_release.graphml")
 #' 
 #' mm <- mixmat(g, "vosonCA_Stance", use_density = FALSE)
 #' }