+++ inst/vosondash/server.R
@@ -38,9 +38,9 @@ shinyServer(function(input, output, session) {
   
   source("server/consoleServer.R", local = TRUE)
   
-  if (is2910) {
-    shinyjs::enable("twitter_semantic_assoc") 
-  }
+  if (is2910) { shinyjs::enable("twitter_semantic_assoc") }
+  
+  if (VOSONDash::isMac()) { shinyjs::enable("macos_font_check") }
   
   # reset collect consoles on startup
   observeEvent(input$sidebar_menu, {